BUG FIXED: DECODER Offset for THEN PART, was not decoded, asummed as integer literal. READY!
NEW BUGS: May be same problem with ELSE PART, more testing to be done.
NEW BUGS: TIME FORMAT LITERALS are not encoded or decoded according to discovery of enclosing byte 157 for number literal and ending byte 63 to differentiate from common number literals.
PENDING: PIDS are the same as controllers? It seems the same. IF so, controllers need support from grammar, encoder and decoder.
More testing to be done: So im gonna build a bytes calculator for each type of literals as a helper tool to programmers and debbuging.